Adding foreground / background color support.
No documentation yet.  Background color is in tight box around text; probably
want to tweak to give some padding.

Since there are upstream bugs in vgfont, for now I'm including vgfont in this
package, patching it locally, and linking to the bundled version.  Once
upstream is patched this can be dropped.